+++ src/backend/joanie/tests/core/test_api_admin_orders.py
@@ -483,6 +483,36 @@ class OrdersAdminApiTestCase(TestCase):
             self.assertEqual(content["count"], 1)
             self.assertEqual(content["results"][0]["id"], str(order.id))
 
+    def test_api_admin_orders_list_filter_by_id(self):
+        """
+        Authenticated admin user should be able to list all existing orders filtered by
+        id
+        """
+        orders = factories.OrderFactory.create_batch(3)
+
+        admin = factories.UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=admin.username, password="password")
+
+        response = self.client.get("/api/v1.0/admin/orders/")
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        content = response.json()
+        self.assertEqual(content["count"], 3)
+
+        response = self.client.get(f"/api/v1.0/admin/orders/?ids={orders[0].id}")
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        content = response.json()
+        self.assertEqual(content["count"], 1)
+        self.assertEqual(content["results"][0]["id"], str(orders[0].id))
+
+        response = self.client.get(
+            f"/api/v1.0/admin/orders/?ids={orders[0].id}&ids={orders[1].id}&ids={orders[1].id}"
+        )
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        content = response.json()
+        self.assertEqual(content["count"], 2)
+        self.assertEqual(content["results"][0]["id"], str(orders[1].id))
+        self.assertEqual(content["results"][1]["id"], str(orders[0].id))
+
     def test_api_admin_orders_course_retrieve(self):
         """An admin user should be able to retrieve a single course order through its id."""