+++ src/backend/joanie/tests/core/test_api_admin_organizations.py
@@ -190,6 +190,37 @@ class OrganizationAdminApiTest(TestCase):
         self.assertEqual(content["count"], 1)
         self.assertEqual(content["results"][0]["title"], "Université")
 
+    def test_admin_api_organization_list_filter_by_id(self):
+        """
+        Staff user should be able to get a paginated list of organizations
+        filtered an id.
+        """
+        admin = factories.UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=admin.username, password="password")
+        organizations = factories.OrganizationFactory.create_batch(3)
+
+        response = self.client.get("/api/v1.0/admin/organizations/")
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        content = response.json()
+        self.assertEqual(content["count"], 3)
+
+        response = self.client.get(
+            f"/api/v1.0/admin/organizations/?ids={organizations[0].id}",
+        )
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        content = response.json()
+        self.assertEqual(content["count"], 1)
+        self.assertEqual(content["results"][0]["id"], str(organizations[0].id))
+
+        response = self.client.get(
+            f"/api/v1.0/admin/organizations/?ids={organizations[0].id}&ids={organizations[1].id}",
+        )
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        content = response.json()
+        self.assertEqual(content["count"], 2)
+        self.assertEqual(content["results"][0]["id"], str(organizations[1].id))
+        self.assertEqual(content["results"][1]["id"], str(organizations[0].id))
+
     def test_admin_api_organization_get(self):
         """
         Staff user should be able to get an organization through its id with detailed